# perf(sql): speed up GROUP BY queries with large result set