--- a/core/src/main/c/linux/files.c
+++ b/core/src/main/c/linux/files.c
@@ -150,14 +150,6 @@ JNIEXPORT jint JNICALL Java_io_questdb_std_Files_getPosixMadvSequential(JNIEnv *
     return POSIX_MADV_SEQUENTIAL;
 }
 
-JNIEXPORT jint JNICALL Java_io_questdb_std_Files_getLinuxMadvPopulateWrite(JNIEnv *e, jclass cls) {
-#if defined(MADV_POPULATE_WRITE)
-    return MADV_POPULATE_WRITE;
-#else
-    return -1;
-#endif
-}
-
 JNIEXPORT jlong JNICALL Java_io_questdb_std_Files_getFileSystemStatus
         (JNIEnv *e, jclass cl, jlong lpszName) {
     struct statfs sb;