--- a/core/src/main/java/io/questdb/PropServerConfiguration.java
+++ b/core/src/main/java/io/questdb/PropServerConfiguration.java
@@ -1199,7 +1199,7 @@ public class PropServerConfiguration implements ServerConfiguration {
             this.cairoGroupByMergeShardQueueCapacity = Numbers.ceilPow2(getInt(properties, env, PropertyKey.CAIRO_SQL_PARALLEL_GROUP_BY_MERGE_QUEUE_CAPACITY, defaultReduceQueueCapacity));
             this.cairoGroupByShardingThreshold = getInt(properties, env, PropertyKey.CAIRO_SQL_PARALLEL_GROUP_BY_SHARDING_THRESHOLD, 100_000);
             this.cairoGroupByPresizeMaxSize = getLong(properties, env, PropertyKey.CAIRO_SQL_PARALLEL_GROUP_BY_PRESIZE_MAX_SIZE, 100_000_000);
-            this.cairoGroupByPresizeMaxHeapSize = getInt(properties, env, PropertyKey.CAIRO_SQL_PARALLEL_GROUP_BY_PRESIZE_MAX_HEAP_SIZE, 100 * Numbers.SIZE_1MB);
+            this.cairoGroupByPresizeMaxHeapSize = getInt(properties, env, PropertyKey.CAIRO_SQL_PARALLEL_GROUP_BY_PRESIZE_MAX_HEAP_SIZE, 200 * Numbers.SIZE_1MB);
             this.cairoPageFrameReduceRowIdListCapacity = Numbers.ceilPow2(getInt(properties, env, PropertyKey.CAIRO_PAGE_FRAME_ROWID_LIST_CAPACITY, 256));
             this.cairoPageFrameReduceColumnListCapacity = Numbers.ceilPow2(getInt(properties, env, PropertyKey.CAIRO_PAGE_FRAME_COLUMN_LIST_CAPACITY, 16));
             final int defaultReduceShardCount = Math.min(sharedWorkerCount, 4);