--- a/core/src/main/java/io/questdb/cairo/CairoConfiguration.java
+++ b/core/src/main/java/io/questdb/cairo/CairoConfiguration.java
@@ -178,6 +178,10 @@ public interface CairoConfiguration {
 
     int getGroupByPoolCapacity();
 
+    long getGroupByPresizeMaxHeapSize();
+
+    long getGroupByPresizeMaxSize();
+
     int getGroupByShardingThreshold();
 
     @NotNull