--- a/core/src/main/java/io/questdb/cairo/CairoConfiguration.java
+++ b/core/src/main/java/io/questdb/cairo/CairoConfiguration.java
@@ -552,6 +552,8 @@ public interface CairoConfiguration {
 
     int getWriterTickRowsCountMod();
 
+    boolean isGroupByPresizeEnabled();
+
     boolean isIOURingEnabled();
 
     boolean isMultiKeyDedupEnabled();