--- a/core/src/main/java/io/questdb/cairo/CairoConfigurationWrapper.java
+++ b/core/src/main/java/io/questdb/cairo/CairoConfigurationWrapper.java
@@ -996,6 +996,11 @@ public class CairoConfigurationWrapper implements CairoConfiguration {
     }
 
     @Override
+    public boolean isGroupByPresizeEnabled() {
+        return delegate.isGroupByPresizeEnabled();
+    }
+
+    @Override
     public boolean isIOURingEnabled() {
         return delegate.isIOURingEnabled();
     }