--- a/core/src/main/java/io/questdb/cairo/map/Map.java
+++ b/core/src/main/java/io/questdb/cairo/map/Map.java
@@ -67,10 +67,6 @@ public interface Map extends Mutable, Closeable, Reopenable {
 
     void restoreInitialCapacity();
 
-    default void setHeapSize(long heapSize) {
-        // no-op
-    }
-
     void setKeyCapacity(int keyCapacity);
 
     long size();