Includes the following changes:
* Prefaults memory used in GROUP BY hash tables and functions (`calloc` calls). This is done by calling `madvise(MADV_POPULATE_WRITE)` in a backward-compatible manner: we hardcode the `LINUX_MADV_POPULATE_WRITE` constant value instead of trying to fetch it from glibc. Thanks to this, on pre-5.14 Linux kernels the `madvise(MADV_POPULATE_WRITE)` call will return EINVAL, so it's a no-op. On newer kernels, prefaulting helps to avoid minor page faults during large hash table accesses - those accesses are random, so fault-ahead doesn't help and the overhead is more than noticeable.
* Adds per-factory statistics for GROUP BY hash tables, including the shards (if they were in use). In many cases, this leads to no `rehash()`/`resize()` calls being made during aggregation/merge when a query is called the second time.
* Tweaks hash functions from `Hash` based on ideas borrowed from Rust compiler's FxHash hash function.
* Also changes hash function's result to 64-bit - to have more entropy in MSBs used for sharding.

A few microbenchmarks.

Before:
```
Benchmark                         (size)  Mode  Cnt    Score    Error  Units
MapReadBenchmark.testHashMap        1000  avgt    3   22.312 ±  0.489  ns/op
MapReadBenchmark.testHashMap       10000  avgt    3   26.617 ±  2.515  ns/op
MapReadBenchmark.testHashMap      100000  avgt    3   53.810 ±  8.596  ns/op
MapReadBenchmark.testHashMap     1000000  avgt    3  277.973 ± 16.110  ns/op
MapReadBenchmark.testOrderedMap     1000  avgt    3   47.701 ±  9.027  ns/op
MapReadBenchmark.testOrderedMap    10000  avgt    3   52.499 ± 31.773  ns/op
MapReadBenchmark.testOrderedMap   100000  avgt    3   59.322 ±  1.008  ns/op
MapReadBenchmark.testOrderedMap  1000000  avgt    3  232.842 ± 26.554  ns/op
```

After:
```
Benchmark                         (size)  Mode  Cnt    Score    Error  Units
MapReadBenchmark.testHashMap        1000  avgt    3   25.515 ±  3.958  ns/op
MapReadBenchmark.testHashMap       10000  avgt    3   29.682 ±  0.278  ns/op
MapReadBenchmark.testHashMap      100000  avgt    3   53.973 ±  3.952  ns/op
MapReadBenchmark.testHashMap     1000000  avgt    3  278.126 ± 45.794  ns/op
MapReadBenchmark.testOrderedMap     1000  avgt    3   27.085 ± 7.102  ns/op
MapReadBenchmark.testOrderedMap    10000  avgt    3   30.054 ± 7.604  ns/op
MapReadBenchmark.testOrderedMap   100000  avgt    3   46.027 ± 7.377  ns/op
MapReadBenchmark.testOrderedMap  1000000  avgt    3  243.714 ± 1.779  ns/op
```

Before:
```
Benchmark                                 (size)  Mode  Cnt    Score    Error  Units
MapReadLongBenchmark.testHashMap            5000  avgt    3    8.190 ±  1.390  ns/op
MapReadLongBenchmark.testHashMap           50000  avgt    3   12.832 ±  2.481  ns/op
MapReadLongBenchmark.testHashMap          500000  avgt    3   62.734 ±  1.911  ns/op
MapReadLongBenchmark.testHashMap         5000000  avgt    3   85.652 ±  2.939  ns/op
MapReadLongBenchmark.testOrderedMap         5000  avgt    3   20.393 ±  1.243  ns/op
MapReadLongBenchmark.testOrderedMap        50000  avgt    3   21.366 ± 11.963  ns/op
MapReadLongBenchmark.testOrderedMap       500000  avgt    3   69.385 ±  6.378  ns/op
MapReadLongBenchmark.testOrderedMap      5000000  avgt    3  151.601 ± 23.570  ns/op
MapReadLongBenchmark.testUnordered16Map     5000  avgt    3   13.934 ±  3.655  ns/op
MapReadLongBenchmark.testUnordered16Map    50000  avgt    3   14.103 ±  1.529  ns/op
MapReadLongBenchmark.testUnordered16Map   500000  avgt    3   39.315 ±  3.733  ns/op
MapReadLongBenchmark.testUnordered16Map  5000000  avgt    3   57.244 ± 14.551  ns/op
MapReadLongBenchmark.testUnordered8Map      5000  avgt    3   15.486 ±  0.923  ns/op
MapReadLongBenchmark.testUnordered8Map     50000  avgt    3   12.962 ±  0.878  ns/op
MapReadLongBenchmark.testUnordered8Map    500000  avgt    3   29.561 ±  2.760  ns/op
MapReadLongBenchmark.testUnordered8Map   5000000  avgt    3   50.963 ±  2.219  ns/op
```

After:
```
Benchmark                                 (size)  Mode  Cnt   Score   Error  Units
MapReadLongBenchmark.testOrderedMap         5000  avgt    3   12.411 ±  0.345  ns/op
MapReadLongBenchmark.testOrderedMap        50000  avgt    3   15.578 ± 12.974  ns/op
MapReadLongBenchmark.testOrderedMap       500000  avgt    3   68.432 ±  7.021  ns/op
MapReadLongBenchmark.testOrderedMap      5000000  avgt    3  145.026 ±  8.722  ns/op
MapReadLongBenchmark.testUnordered16Map     5000  avgt    3    8.476 ± 0.567  ns/op
MapReadLongBenchmark.testUnordered16Map    50000  avgt    3    9.160 ± 0.383  ns/op
MapReadLongBenchmark.testUnordered16Map   500000  avgt    3   34.675 ± 1.831  ns/op
MapReadLongBenchmark.testUnordered16Map  5000000  avgt    3   43.018 ± 4.532  ns/op
MapReadLongBenchmark.testUnordered8Map      5000  avgt    3    7.865 ± 0.170  ns/op
MapReadLongBenchmark.testUnordered8Map     50000  avgt    3    8.174 ± 1.278  ns/op
MapReadLongBenchmark.testUnordered8Map    500000  avgt    3   26.619 ± 0.590  ns/op
MapReadLongBenchmark.testUnordered8Map   5000000  avgt    3   40.777 ± 0.791  ns/op
```