@ideoma 2024-01-25
#### [PR Coverage check]

:heart_eyes: **pass** : 96 / 98 (97.96%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/std/Files.java|2|3|66.67%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByAllocator.java|15|16|93.75%|
|:large_blue_circle:|io/questdb/std/DirectIntList.java|6|6|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/Unordered8Map.java|4|4|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByLongHashSet.java|10|10|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByLong128HashSet.java|11|11|100.00%|
|:large_blue_circle:|io/questdb/std/Unsafe.java|15|15|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/Unordered16Map.java|4|4|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByIntHashSet.java|10|10|100.00%|
|:large_blue_circle:|io/questdb/std/FilesFacadeImpl.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/OrderedMap.java|5|5|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByLong256HashSet.java|13|13|100.00%|