@ideoma 2024-01-26
#### [PR Coverage check]

:heart_eyes: **pass** : 298 / 307 (97.07%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/cairo/map/Map.java|0|1|00.00%|
|:large_blue_circle:|io/questdb/cairo/CairoConfigurationWrapper.java|0|2|00.00%|
|:large_blue_circle:|io/questdb/std/Files.java|3|4|75.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByAllocator.java|18|19|94.74%|
|:large_blue_circle:|io/questdb/cairo/map/Unordered8Map.java|19|20|95.00%|
|:large_blue_circle:|io/questdb/cairo/map/Unordered16Map.java|21|22|95.45%|
|:large_blue_circle:|io/questdb/cairo/map/OrderedMap.java|40|42|95.24%|
|:large_blue_circle:|io/questdb/PropServerConfiguration.java|5|5|100.00%|
|:large_blue_circle:|io/questdb/std/DirectIntList.java|8|8|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByLongHashSet.java|10|10|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByLong128HashSet.java|11|11|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncGroupByAtom.java|116|116|100.00%|
|:large_blue_circle:|io/questdb/cairo/DefaultCairoConfiguration.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/std/Unsafe.java|10|10|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByIntHashSet.java|10|10|100.00%|
|:large_blue_circle:|io/questdb/std/FilesFacadeImpl.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/PropertyKey.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncGroupByRecordCursorFactory.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncGroupByRecordCursor.java|3|3|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByLong256HashSet.java|13|13|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/MapFactory.java|4|4|100.00%|